Changes the color triggering of buttons
Adds a new function toggleColor(), and uses setTimeout for that.
Gets rid of mouseup triggers in buttonsFeedback(), which ended
up making mobile work. Also, toggling colors between opacities
0.3 and 1.0 was an issue, because Firefox at least was changing
the css value to rgb() instead of rgba() when opacity was changed
in javascript to 1.0. Apparently it figured that we didn't need
the 'a' in 'rgba' anymore. I WAS doing some string slicing with
the anticipation that 'rgba' was present as was the alpha value
at the end of the color value. Oh well.